+++ src/ui_utils/online_message.py
@@ -2,9 +2,9 @@ from mods_base import get_pc
 
 __all__: tuple[str, ...] = (
     "hide_blocking_message",
-    "hide_message",
+    "hide_coop_message",
     "show_blocking_message",
-    "show_message",
+    "show_coop_message",
 )
 
 
@@ -12,9 +12,9 @@ def show_blocking_message(msg: str, reason: str | None = None) -> None:
     """
     Displays a blocking message with the given text.
 
-    This message will stay until it is explicitly hidden, see `hide_blocking_message`.
+    This message blocks all user input until it is hidden.
 
-    This message will block any input until it is hidden.
+    This message will stay until it is explicitly hidden, see `hide_blocking_message`.
 
     Args:
         msg: The message to display.
@@ -38,11 +38,11 @@ def hide_blocking_message() -> None:
     msg_movie.HideBlocking()
 
 
-def show_message(msg: str) -> None:
+def show_coop_message(msg: str) -> None:
     """
-    Displays a message on the left side of the screen.
+    Displays a short message on the left of the screen, like those used when coop players join.
 
-    This message will stay until it is explicitly hidden, see `hide_message`.
+    This message will stay until it is explicitly hidden, see `hide_coop_message`.
 
     Args:
         msg: The message to display.
@@ -53,8 +53,8 @@ def show_message(msg: str) -> None:
     msg_movie.DisplayMessage(msg)
 
 
-def hide_message() -> None:
-    """Hides the currently displayed message in the online message movie."""
+def hide_coop_message() -> None:
+    """Hides the currently displayed coop message, if any."""
     if (msg_movie := get_pc().GetOnlineMessageMovie()) is None:
         return